+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### [unreleased]
+* add bypass for WordPress 5.1 WSOD Shutdown Handler
 
 #### 2.2.0 / 2018-11-25
 * use composer for dependencies and autoloader